refactor: Move definitions from workspace to npm
