+++ packages/preload/src/modules/npm.ts
@@ -6,6 +6,10 @@ import { invoke } from './invoke';
  * @param {string} path - The file system path of the project where dependencies should be installed.
  * @returns {Promise<void>} - A promise that resolves when the installation is complete.
  */
-export async function install(path: string): Promise<void> {
-  await invoke('npm.install', path);
+export async function install(path: string, _package?: string): Promise<void> {
+  await invoke('npm.install', path, _package);
+}
+
+export async function packageOutdated(_path: string, _package: string): Promise<boolean> {
+  return invoke('npm.packageOutdated', _path, _package);
 }
