+++ packages/shared/types/ipc.ts
@@ -9,8 +9,6 @@ export interface Ipc {
   'electron.openExternal': (url: string) => Promise<void>;
   'inspector.start': () => Promise<number>;
   'bin.install': () => Promise<void>;
-  'bin.installNpmPackages': (projectPath: string, packageName?: string) => Promise<void>;
-  'bin.npmPackageOutdated': (projectPath: string, packageName: string) => Promise<boolean>;
   'bin.code': (path: string) => Promise<void>;
   'cli.init': (path: string, repo?: string) => Promise<void>;
   'cli.start': (path: string) => Promise<void>;
@@ -18,5 +16,6 @@ export interface Ipc {
   'analytics.track': (event: string, data?: Record<string, any>) => void;
   'analytics.identify': (userId: string, traits?: Record<string, any>) => void;
   'analytics.getAnonymousId': () => Promise<string>;
-  'npm.install': (path: string) => Promise<void>;
+  'npm.install': (path: string, packageName?: string) => Promise<void>;
+  'npm.packageOutdated': (path: string, packageName: string) => Promise<boolean>;
 }
